==> notebooks/01_GameSpy_Scraping_jeuxvideo.com.ipynb <==
# Import des données depuis la page de recherche regroupant tous les tests

from bs4 import BeautifulSoup
import requests
import pandas as pd

# Récupération des liens amenant à chaque fiche jeu

links=[]

#Boucle pour récupérer les liens dans les pages de listing des tests de jeuxvideo.com
for x in range(1,933): 
    url="https://www.jeuxvideo.com/tests/?p=" 
    r=requests.get(url+str(x)) #Il y a 932 pages de tests, x vient complèter l'url pour scraper toutes les pages
    soup=BeautifulSoup(r.content,"html.parser")
    content=soup.find_all("div",class_="container__3zUztO iconPlayHover") #Récupération du bloc principal de la page

    #Boucle pour récupérer, parmi les éléments du bloc principal de la page scrapée, 
    #les extensions des liens amenant vers la page test de chaque jeu
    for property in content:
        link=property.find('a', {"class": "bonusLabel__3bIEJW gamePageLink__25_ek7"}, href=True)
        if link!=None:
            links.append(link["href"])
        
print(links)

#Affichage du nombre d'extensions de liens récupérés
print(len(links))

# VERSION 1 : import des données depuis chaque fiche jeu via boucle for

#Création d'une liste
games_jv3=[]

#Boucle qui pour chaque lien des pages test récupérées va permettre de chercher le nom du jeu, les plateformes,
#le nombre d'avis par plateforme, la note joueurs par plateforme, la note des journalistes et la date de lancement
for y in links:
        url2 = "https://www.jeuxvideo.com"
        r2 = requests.get(url2 + y) #chaque extention de lien est appelée pour complèter le host name de jeuxvideo.com

        soup2 = BeautifulSoup(r2.content, "html.parser")
        content2 = soup2.find_all("div", class_="layout layout--classic js-layout")
        
        names = soup2.find("h1",class_="breadcrumb__item").text #récupération des noms du jeu vidéo
        
        #récupération des plateformes sur lesquelles sont sorties le jeu    
        plateforms = soup2.find_all("span",class_="userOpinionsPerPlatform__platformLabel") 
        
        #récupération des avis par plateforme    
        reviews = soup2.find_all("span",class_="userOpinionsPerPlatform__ratingCount")
        
        #récupération des notes des joueurs par plateforme
        players_rates = soup2.find_all("span",class_="userOpinionsPerPlatform__averageRating")
        
        #récupération de la note journalistes si existante
        zone_rate_journalist=soup2.find("div",class_="gameCharacteristicsMain__gauge")
        if zone_rate_journalist == None:
            rate_journalist = None
        else:
            rate_journalist = soup2.find("text",class_="gameCharacteristicsMain__gaugeText").text
        
        #récupération de la date de lancement du jeu si présente sur la page
        zone_launch_date = soup2.find("div",class_="gameCharacteristicsMain__releaseDate")
        if zone_launch_date == None:
            launch_date = None
        else:
            launch_date = soup2.find("div",class_="gameCharacteristicsMain__releaseDate").text
        
        #création d'un dictionnaire avec les données récupérées
        property_info3={"names":names,
                        "plateforms":plateforms,
                        "reviews":reviews,
                        "players_rates":players_rates,
                        "journalist_rate":rate_journalist,
                       "launch_date":launch_date}
        
        #Enrichissement de la liste avec le dictionnaire créé
        games_jv3.append(property_info3)

#Création d'un dataframe avec ajustement de la date de lancement        
df=pd.DataFrame(games_jv3)
df["launch_date"]=df["launch_date"].apply(lambda x: x[12:]).str.strip()
df

# VERSION 2 : import des données depuis chaque fiche jeu via déf fonction

#Fonction qui prend en paramètre l'url d'une page test de jeu vidéo et qui permet de retourner le nom du jeu, les plateformes,
#le nombre d'avis par plateformes, la note joueurs par plateforme, la note des journalistique et la date de lancement
def extract_data(url_test):
        names = url_test.find("h1",class_="breadcrumb__item").text
        
            
        plateforms = url_test.find_all("span",class_="userOpinionsPerPlatform__platformLabel")
        
            
        reviews = url_test.find_all("span",class_="userOpinionsPerPlatform__ratingCount")
        
        
        players_rates = url_test.find_all("span",class_="userOpinionsPerPlatform__averageRating")
        
        zone_rate_journalist = url_test.find("div",class_="gameCharacteristicsMain__gauge")
        if zone_rate_journalist == None:
            rate_journalist = None
        else:
            rate_journalist = url_test.find("text",class_="gameCharacteristicsMain__gaugeText").text
            
        zone_launch_date = url_test.find("div",class_="gameCharacteristicsMain__releaseDate")
        if zone_launch_date == None:
            launch_date = None
        else:
            launch_date = url_test.find("div",class_="gameCharacteristicsMain__releaseDate").text
    
        property_info3={"names":names,
                        "plateforms":plateforms,
                        "reviews":reviews,
                        "players_rates":players_rates,
                        "journalist_rate":rate_journalist,
                       "launch_date":launch_date}
    
        return property_info3

#Création d'une liste
data_full = []
#Création d'une boucle qui va parcourir toutes les urls des pages de test récupérées sur jeuxvideo.com
for url in links:
    url_scrap = "https://www.jeuxvideo.com"+url
    url_page = requests.get(url_scrap)
    parsed_url_page = BeautifulSoup(url_page.content, 'html.parser')
    #Alimentation de la liste avec les résultats obtenus de la fonction extract_data pour chaque url
    data_full.append(extract_data(parsed_url_page)) 

#Création d'un dataframe
df=pd.DataFrame.from_dict(data_full)
df

# Nettoyage des colonnes Plateformes, Avis et Notes joueurs

df1=df

#Boucle permettant de supprimer les crochets pour chaque valeur à l'intérieur des listes de plateformes
clean_plateforms=[]
for p in df1["plateforms"]:
    sub_clean=[]
    for x in p:
        sub_clean.append(x.text)
    clean_plateforms.append(sub_clean)
#Création d'une nouvelle colonne avec les valeurs nettoyées
df1["clean_plateforms"]=clean_plateforms

#Boucle permettant de supprimer les crochets pour chaque valeur à l'intérieur des listes d'avis
clean_reviews=[]
for p in df1["reviews"]:
    sub_clean=[]
    for x in p:
        sub_clean.append(x.text)
    clean_reviews.append(sub_clean)
#Création d'une nouvelle colonne avec les valeurs nettoyées    
df1["clean_reviews"]=clean_reviews

#Boucle permettant de supprimer les crochets pour chaque valeur à l'intérieur des listes de notes de joueurs
clean_rates=[]
for p in df1["players_rates"]:
    sub_clean=[]
    for x in p:
        sub_clean.append(x.text)
    clean_rates.append(sub_clean)
#Création d'une nouvelle colonne avec les valeurs nettoyées                
df1["clean_players_rates"]=clean_rates

#Suppression des anciennes colonnes    
df1.pop("plateforms")
df1.pop("reviews")
df1.pop("players_rates")
df1

# Remise en ordre des notes par rapport aux avis

#Copie de la nouvelle colonne notes joueurs dans une variable appellée tab
tab=df1["clean_players_rates"]

#Boucle permettant, pour une même ligne, d'avoir autant de notes joueurs que de nombres d'avis
#Une valeur "0 avis" n'avait pas de correspondance note joueur --> ajout de 0 dans les notes joueurs pour avoir le même nombre d'items
for i in range(len(tab)):
    if len(df1["clean_reviews"][i])>len(tab[i]):
        diff=len(df1["clean_reviews"][i])-len(tab[i])
        tab[i].extend(0 for x in range(diff))

#Boucle permettant de créer une nouvelle colonne avec le bon nombre d'items avis/notes joueurs par ligne et une remise à zéro
#Cette colonne sera utile pour remettre les notes joueurs en face du nombre d'avis laissés
tab2=[]
for i in tab:
    sub_tab2=[]
    for j in range(len(i)):
        sub_tab2.append(0)
    tab2.append(sub_tab2)
    
df1["players_rates_ordered"]=tab2
df1

#Création de 3 listes sur la base des colonnes nombre d'avis, des notes joueurs dans le désordre 
#et de la dernière colonne avec le bon nombre d'items avis/notes mais que des valeurs à zéros 
list1=list(df1["clean_reviews"])
list2=list(df1["clean_players_rates"])
list3=list(df1["players_rates_ordered"])

#Boucle qui parcours les listes du nombre d'avis et des notes joueurs et qui vient mettre dans l'ordre les notes joueurs
#avec un 0 si le nombre d'avis est à 0 ou avec la note correspondant aux nombre d'avis sinon
for i in range(len(list1)):
    index=0
    for j in range(len(list1[i])):
            if "0" in list1[i][j]:
                list3[i][j]=0
            else:
                list3[i][j]=list2[i][index]
                index+=1

df1["players_rates_ordered"]=list3
df1.pop("clean_players_rates")
df1

# Explode du dataset pour avoir une entrée Jeu/Plateforme

#Eclatement des listes de plateformes, du nombre d'avis et de notes joueurs sur différentes lignes
#pour garder des lignes avec une entrée jeu/plateforme, un nombre d'avis et une note sur la combinaison jeu/plateforme
data_jv_final=df1.set_index(['names',"journalist_rate","launch_date"]).apply(pd.Series.explode).reset_index()
data_jv_final

# Création d'une colonne liée à l'année de lancement du jeu

#Création d'une colonne avec uniquement l'année de lancement
data_jv_final["year"]=data_jv_final["launch_date"].apply(lambda x: x[-4:])
data_jv_final

# Filtrage des jeux ne contenant pas le nom de la plateforme dans le nom du jeu

#Filtre du dataset pour exclure les jeux qui comportent le nom de la plateforme dans le nom récupéré 
#et qui ne pourront pas être matchés avec le dataset initial. Exemple : "Pokémon sur Switch" ou "Animal Crossing sur PC"
data_jv_final=data_jv_final[data_jv_final["names"].str.contains(" sur ")==False]
data_jv_final

# Export du dataset final

data_jv_final.to_csv("../data/01_GameSpy_Scrapping_jv.csv",sep=',')
